test: [alb] consolidate per-axis race/panic tests and lift integration prom stubs
